+++ apps/api/api/delete.go
@@ -5,16 +5,21 @@ import (
 	"connectrpc.com/connect"
 	"context"
 	"fmt"
+	"github.com/getsentry/sentry-go"
 	"go.dot.industries/brease/auth"
+	"go.uber.org/zap"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
 func (b *BreaseHandler) DeleteRule(ctx context.Context, c *connect.Request[contextv1.DeleteRuleRequest]) (*connect.Response[emptypb.Empty], error) {
 	orgID := auth.CtxString(ctx, auth.ContextOrgKey)
-	if !auth.HasPermission(ctx, auth.PermissionWrite) {
+	if !auth.HasPermission(ctx, auth.PermissionCreateRule) {
 		return nil, connect.NewError(connect.CodePermissionDenied, fmt.Errorf("permission denied"))
 	}
-	_ = b.db.RemoveRule(ctx, orgID, c.Msg.ContextId, c.Msg.RuleId)
-	// we don't expose whether we succeeded
+	err := b.db.RemoveRule(ctx, orgID, c.Msg.ContextId, c.Msg.RuleId)
+	if err != nil {
+		sentry.CaptureException(err)
+		b.logger.Error("Failed to delete rule", zap.Error(err))
+	}
 	return connect.NewResponse(&emptypb.Empty{}), nil
 }