+++ apps/api/storage/buntdb/db.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"go.dot.industries/brease/env"
 	"google.golang.org/protobuf/proto"
 	"sync"
 
@@ -20,7 +21,7 @@ type Options struct {
 	Logger *zap.Logger
 }
 
-func NewDatabase(opts Options) (storage.Database, error) {
+func NewDatabase(opts Options) (*Container, error) {
 	if opts.Path == "" {
 		opts.Path = ":memory:"
 	}
@@ -32,7 +33,7 @@ func NewDatabase(opts Options) (storage.Database, error) {
 		return nil, err
 	}
 
-	b := &buntdbContainer{
+	b := &Container{
 		db:     db,
 		logger: opts.Logger,
 		rulePool: sync.Pool{
@@ -47,17 +48,17 @@ func NewDatabase(opts Options) (storage.Database, error) {
 	return b, nil
 }
 
-type buntdbContainer struct {
+type Container struct {
 	db       *buntdb.DB
 	logger   *zap.Logger
 	rulePool sync.Pool
 }
 
-func (b *buntdbContainer) Close() error {
+func (b *Container) Close() error {
 	return b.db.Close()
 }
 
-func (b *buntdbContainer) AddRule(_ context.Context, ownerID string, contextID string, rule *rulev1.Rule) (*rulev1.VersionedRule, error) {
+func (b *Container) AddRule(_ context.Context, ownerID string, contextID string, rule *rulev1.Rule) (*rulev1.VersionedRule, error) {
 	vRule := &rulev1.VersionedRule{
 		Id:          rule.Id,
 		Version:     1,
@@ -67,11 +68,15 @@ func (b *buntdbContainer) AddRule(_ context.Context, ownerID string, contextID s
 	}
 	rk := storage.RuleKey(ownerID, contextID, vRule.Id)
 	vk := storage.VersionKey(ownerID, contextID, vRule.Id, vRule.Version)
-
 	ruleJSON, err := proto.Marshal(vRule)
 	if err != nil {
 		return nil, err
 	}
+
+	if env.IsDebug() {
+		b.logger.Debug("Adding rule", zap.String("ruleKey", rk), zap.String("versionKey", vk), zap.String("ruleJSON", string(ruleJSON)))
+	}
+
 	err = b.db.Update(func(tx *buntdb.Tx) error {
 		_, _, txErr := tx.Set(rk, vk, nil)
 		if txErr != nil {
@@ -87,9 +92,19 @@ func (b *buntdbContainer) AddRule(_ context.Context, ownerID string, contextID s
 	return vRule, err
 }
 
-func (b *buntdbContainer) Rules(_ context.Context, ownerID string, contextID string, pageSize int, pageToken string) (rules []*rulev1.VersionedRule, err error) {
+func (b *Container) Rules(_ context.Context, ownerID string, contextID string, pageSize int, pageToken string) (rules []*rulev1.VersionedRule, err error) {
 	rkSearch := storage.RuleKey(ownerID, contextID, "*")
 
+	if env.IsDebug() {
+		_ = b.db.View(func(tx *buntdb.Tx) error {
+			return tx.AscendKeys(rkSearch, func(key, val string) bool {
+				if storage.IsVersionKey(val) {
+					b.logger.Debug("rule key", zap.String("key", key), zap.String("val", val))
+				}
+				return true
+			})
+		})
+	}
 	var ruleKeys []string
 	err = b.db.View(func(tx *buntdb.Tx) error {
 		return tx.AscendKeys(rkSearch, func(key, val string) bool {
@@ -108,21 +123,28 @@ func (b *buntdbContainer) Rules(_ context.Context, ownerID string, contextID str
 				return vErr
 			}
 
-			rule := b.rulePool.Get().(*rulev1.VersionedRule)
+			rule := &rulev1.VersionedRule{}
+
 			umErr := proto.Unmarshal([]byte(latestVersionData), rule)
 			if umErr != nil {
 				return umErr
 			}
+			if env.IsDebug() {
+				b.logger.Debug("rule version", zap.String("versionKey", vk), zap.String("ruleID", rule.Id), zap.Any("version", rule), zap.String("latestVersion", latestVersionData))
+			}
 			rules = append(rules, rule)
-			b.rulePool.Put(rule)
 		}
 		return nil
 	})
 
+	if env.IsDebug() {
+		b.logger.Debug("Rules", zap.Any("rules", rules))
+	}
+
 	return
 }
 
-func (b *buntdbContainer) RuleVersions(_ context.Context, ownerID string, contextID string, ruleID string, pageSize int, pageToken string) (rules []*rulev1.VersionedRule, err error) {
+func (b *Container) RuleVersions(_ context.Context, ownerID string, contextID string, ruleID string, pageSize int, pageToken string) (rules []*rulev1.VersionedRule, err error) {
 	vkSearch := storage.RuleKey(ownerID, contextID, ruleID) + ":v*"
 	var ruleVersions []string
 	err = b.db.View(func(tx *buntdb.Tx) error {
@@ -134,13 +156,13 @@ func (b *buntdbContainer) RuleVersions(_ context.Context, ownerID string, contex
 
 	err = b.db.View(func(tx *buntdb.Tx) error {
 		for _, versionData := range ruleVersions {
-			rule := b.rulePool.Get().(*rulev1.VersionedRule)
+			rule := &rulev1.VersionedRule{}
+
 			umErr := proto.Unmarshal([]byte(versionData), rule)
 			if umErr != nil {
 				return umErr
 			}
 			rules = append(rules, rule)
-			b.rulePool.Put(rule)
 		}
 		return nil
 	})
@@ -148,7 +170,7 @@ func (b *buntdbContainer) RuleVersions(_ context.Context, ownerID string, contex
 	return
 }
 
-func (b *buntdbContainer) RemoveRule(_ context.Context, ownerID string, contextID string, ruleID string) error {
+func (b *Container) RemoveRule(_ context.Context, ownerID string, contextID string, ruleID string) error {
 	rk := storage.RuleKey(ownerID, contextID, ruleID)
 	vkSearch := storage.RuleKey(ownerID, contextID, ruleID) + ":v*"
 	var ruleVersions []string
@@ -179,7 +201,7 @@ func (b *buntdbContainer) RemoveRule(_ context.Context, ownerID string, contextI
 	})
 }
 
-func (b *buntdbContainer) ruleLatestVersion(ruleKey string) (version uint64, err error) {
+func (b *Container) ruleLatestVersion(ruleKey string) (version uint64, err error) {
 	version = 0
 	err = b.db.View(func(tx *buntdb.Tx) error {
 		vk, gErr := tx.Get(ruleKey)
@@ -192,7 +214,7 @@ func (b *buntdbContainer) ruleLatestVersion(ruleKey string) (version uint64, err
 	return
 }
 
-func (b *buntdbContainer) ReplaceRule(_ context.Context, ownerID string, contextID string, rule *rulev1.Rule) (*rulev1.VersionedRule, error) {
+func (b *Container) ReplaceRule(_ context.Context, ownerID string, contextID string, rule *rulev1.Rule) (*rulev1.VersionedRule, error) {
 	rk := storage.RuleKey(ownerID, contextID, rule.Id)
 	currentVersion, err := b.ruleLatestVersion(rk)
 	if err != nil {
@@ -226,7 +248,7 @@ func (b *buntdbContainer) ReplaceRule(_ context.Context, ownerID string, context
 	return vRule, err
 }
 
-func (b *buntdbContainer) Exists(_ context.Context, ownerID string, contextID string, ruleID string) (exists bool, err error) {
+func (b *Container) Exists(_ context.Context, ownerID string, contextID string, ruleID string) (exists bool, err error) {
 	err = b.db.View(func(tx *buntdb.Tx) error {
 		_, ierr := tx.Get(storage.RuleKey(ownerID, contextID, ruleID), true)
 		switch {
@@ -242,7 +264,7 @@ func (b *buntdbContainer) Exists(_ context.Context, ownerID string, contextID st
 	return
 }
 
-func (b *buntdbContainer) SaveAccessToken(c context.Context, ownerID string, tokenPair *authv1.TokenPair) error {
+func (b *Container) SaveAccessToken(c context.Context, ownerID string, tokenPair *authv1.TokenPair) error {
 	atKey := fmt.Sprintf("access:%s", ownerID)
 
 	tokens, err := b.GetAccessTokens(c, ownerID)
@@ -262,7 +284,7 @@ func (b *buntdbContainer) SaveAccessToken(c context.Context, ownerID string, tok
 	})
 }
 
-func (b *buntdbContainer) GetAccessTokens(_ context.Context, ownerID string) (tp []*authv1.TokenPair, err error) {
+func (b *Container) GetAccessTokens(_ context.Context, ownerID string) (tp []*authv1.TokenPair, err error) {
 	atKey := fmt.Sprintf("access:%s", ownerID)
 
 	val := ""
@@ -290,7 +312,7 @@ func (b *buntdbContainer) GetAccessTokens(_ context.Context, ownerID string) (tp
 	return tp, nil
 }
 
-func (b *buntdbContainer) createIndices() {
+func (b *Container) createIndices() {
 	err := b.db.CreateIndex("contexts", storage.ContextKey("*", "*"))
 	if err != nil {
 		panic(err)
@@ -300,4 +322,44 @@ func (b *buntdbContainer) createIndices() {
 	if err != nil {
 		panic(err)
 	}
+}
+
+func (b *Container) GetObjectSchema(ctx context.Context, ownerID string, contextID string) (string, error) {
+	ck := storage.ContextSchemaKey(ownerID, contextID)
+	schema := ""
+	err := b.db.View(func(tx *buntdb.Tx) error {
+		sch, e := tx.Get(ck)
+		if e != nil {
+			return e
+		}
+		schema = sch
+		return nil
+	})
+	if err != nil {
+		return "", err
+	}
+
+	return schema, nil
+}
+
+func (b *Container) ReplaceObjectSchema(ctx context.Context, ownerID string, contextID string, schema string) error {
+	ck := storage.ContextSchemaKey(ownerID, contextID)
+	return b.db.Update(func(tx *buntdb.Tx) error {
+		_, _, e := tx.Set(ck, schema, nil)
+		return e
+	})
+}
+
+func (b *Container) ListContexts(ctx context.Context, ownerID string) ([]string, error) {
+	var contexts []string
+	err := b.db.View(func(tx *buntdb.Tx) error {
+		return tx.AscendKeys(storage.ContextKey(ownerID, "*"), func(key, val string) bool {
+			contexts = append(contexts, key)
+			return true
+		})
+	})
+	if err != nil {
+		return nil, err
+	}
+	return contexts, nil
 }