+++ apps/api/storage/interface.go
@@ -16,4 +16,7 @@ type Database interface {
 	RuleVersions(ctx context.Context, ownerID string, contextID string, ruleID string, pageSize int, pageToken string) ([]*rulev1.VersionedRule, error)
 	SaveAccessToken(ctx context.Context, ownerID string, tokenPair *authv1.TokenPair) error
 	GetAccessTokens(ctx context.Context, ownerID string) ([]*authv1.TokenPair, error)
+	GetObjectSchema(ctx context.Context, ownerID string, contextID string) (string, error)
+	ReplaceObjectSchema(ctx context.Context, ownerID string, contextID string, schema string) error
+	ListContexts(ctx context.Context, ownerID string) ([]string, error)
 }