+++ packages/core/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @brease/core
 
+## 0.5.0
+
+### Minor Changes
+
+- upgrade to latest connectrpc
+
+## 0.4.0
+
+### Minor Changes
+
+- 9adb2c0: Add sequence number to rule object
+
+## 0.3.0
+
+### Minor Changes
+
+- 2f7d9f7: add client side rule decoders
+
+## 0.2.3
+
+### Patch Changes
+
+- cfc78a5: replace jsonpath
+
+## 0.2.2
+
+### Patch Changes
+
+- e596607: package.json updates
+
 ## 0.2.1
 
 ### Patch Changes