+++ packages/react/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @brease/react
 
+## 0.2.0
+
+### Minor Changes
+
+- upgrade to latest connectrpc
+
+### Patch Changes
+
+- Updated dependencies
+  - @brease/core@0.5.0
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [9adb2c0]
+  - @brease/core@0.4.0
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [2f7d9f7]
+  - @brease/core@0.3.0
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [cfc78a5]
+  - @brease/core@0.2.3
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [e596607]
+  - @brease/core@0.2.2
+
 ## 0.1.3
 
 ### Patch Changes